+++ x/erc20/keeper/expected_keeper_innertx.go
@@ -1,0 +1,15 @@
+package keeper
+
+import (
+	sdk "github.com/okex/exchain/libs/cosmos-sdk/types"
+	evmtypes "github.com/okex/exchain/x/evm/types"
+)
+
+type EvmKeeper interface {
+	GetChainConfig(ctx sdk.Context) (evmtypes.ChainConfig, bool)
+	GenerateCSDBParams() evmtypes.CommitStateDBParams
+	GetParams(ctx sdk.Context) evmtypes.Params
+	AddInnerTx(...interface{})
+	AddContract(...interface{})
+	UpdateInnerBlockData(...interface{})
+}
